Modified function with use of lists and indexes
Insted of copying the parameter and using it continuosly, lists and
indexes are now used to reference the arguments. New function to
initialize and allocate the arguments for quads creation is implemented.

Multithreading is now used for efficient calculations of quads.
Further small enhancements are done to modularise the code.