Main structure only keeps intermediate values for references
Until now, the params structure also had an alternative column for the
intermediate calculations for the query table. However, for the query,
we don't need to store all these intermediate values! We just
calculate them and in the same thread we can immediately try matching
them with the reference KD-tree and discard them.

So the parameters have been renamed to avoid confussion and the extra
structure elements have been removed. Also, to allow easy and modular
reading of the input tables, a new function was defined that will
check if indeed three columns have been input and to set them to the
proper pointers.