The C and D vertices are not also separated by their angle
Until now, we were using the Cx<Dx criteria to distinguish between C
and D. However, this could cause degeneracies.

With this commit, I corrected the criteria to select C from D using a
similar smaller angle-based approach that we used for identifying A
and B. Also, a condition was added so when the angles are too close,
the quad will not be used at all.

Also, I done some further cleaning up to help in the readability and
clarity.