Stars that go in a quad don't have similar magnitudes
The order of brightness is an important selection factor in our
implementation. However, due to noise or small measurement
systematics, it may happen that two stars that have close brightnesses
may be confused in different measurements/images.

Therefore with this commit, the operation of selecting vertices for a
quad now goes through an iterative process to only select vertices
that have a magnitude that is sufficiently from each other (with the
new 'magnitude_error' argument). If no quad vertices with sufficient
magnitude differences can be defined, then the quad is marked as
non-existant.