Matching with X-Y catalog done, but similar quads not made
The matching of the quads is now done and it seems to work nicely (the
matched quads do mostly look very similar). However, they don't
correspond to the same group of stars! The problem seems to be that
there actually aren't any similarly selected quads (using same stars!).

I defined some new functions to create a region file for matched quads
to be able to visually compare them along with some further polishing
up to make the code easier to read and to test.

Currently it only builds 1 quad in a 2 by 2 grid structure and I am
using that to see why different stars are being selected, even though
we are only using the brightest stars.