All possible quads around a star are now found (not complete)
Until now, we were only selecting one quad per bright star. But that
is too restrictive and small differences between the reference and
query catalog resulted in not matching the proper quad.

So with this commit, for each bright star, we now build all possible
quads based on the stars nearby, then reject the ones with magnitudes
and geometric shapes that are too similar and only calculate the hash
for the rest.

Currently this isn't yet complete because the hashes for each bright
star are not yet merged into one table for a thread, and later into
one final table. I'll work on that tomorrow.