Working on merging the quads from each star into one list
The quads have been created for each star in each thread, but they are
separate lists of datasets! The preparations have been made to merge
the arrays of datasets into one array at the end of the thread (after
parsing all the bright stars in that thread.

Also, the allocation of the arrays of datasets is now done based on
need (for the query catalog, we don't need to keep the hashes, only
the indexs).