Seems to work: actually matching quads can be found
After trying the script of a known matching quad, I found the problem:
we weren't defining the X and Y axis! The points A, B, C and D were
being found uniquely, but because the X and Y axises weren't, the
first two and second two hash elements were flipped! I kept the parts
related to checking and testing as commented regions to help in future
debuggings.

Also, because the we actually have the different magnitude ratios as
new hash dimensions, magnitude error is redundant. Because, if one of
the B, C or D stars has a magnitude too close to A, then that hash
element will be very near 0, but that is no problem any way.