New problem: apparently same stars are not being used
Apparently, the raw quad hashes are being built nicely, and sometimes
they give a good match. However, in many cases they don't actually
match! After trying to look into it, I feel the problem is because
some quads are not being built in some catalogs. So I have temporarily
disabled the gridding feature to try to build every single possible
quad.

But I have still not fully checked this scenario to confirm that when
a query quad is incorrectly matched, do we indeed have its proper quad
in the k-d tree output?