Matched quads ready for further processing
We still have problems with not finding similar quads in the Gaia
catalog and the image-based catalog. So for now, I just used the
RA/Dec positions of the same image catalog to match with the same
points in the X/Y coordinates and the matches seemed reasonable
(although there are still a few slightly bad matches), but its enough
to get going to later steps now and come back to the problem of
finding same quads later.

With this commit, the matching quad vertices (identified as A, B, C or
D) are kept in arrays, as well as an array to keep the index of the
quad in the matching catalog and they are now simply printed on the
terminal, ready to be used....