Final theta and scale values wrong for some quads
After finally matching the query-quads with the reference-quads and
calculating the value of fo raotation and scaling between them, most of
the values for the quads are fine and approximate na dcan finally be
used to make the CD matrix.

But some values are still erronous and have an opposite sign than other
values for the same quad.